requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-telegram-bot>=21.0
anthropic>=0.40.0
google-genai>=1.0.0
groq>=0.11.0
fastembed>=0.3.0
jinja2>=3.1.0
python-multipart>=0.0.9
httpx>=0.27.0
pgvector==0.2.4
psycopg2-binary==2.9.9
alembic==1.13.1
requests==2.31.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0
rapidfuzz>=3.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
